feat!: allow partial patient and group export (#29)
* feat: allow partial patient and group export (#27)

* chore: update fhir-works-on-aws-interface version

Co-authored-by: Yanyu Zheng <yz2690@columbia.edu>

BREAKING CHANGE: `verifyAccessToken` now allows bulk export requests when the user has read permissions on SOME of the requested resource types instead of requiring permissions on ALL of them. This allows partial exports scoped to the user's permissions. Clients are expected to call `getAllowedResourceTypesForOperation` to scope down the resource types to export.